PN-3496 amount Integer

+++ src/main/java/it/pagopa/pn/deliverypush/dto/timeline/details/NotificationPaidDetails.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class NotificationPaidDetails implements RecipientRelatedTimelineElementDetails {
     private int recIndex;
     private String recipientType;
-    private long amount;
+    private Integer amount;
     private String creditorTaxId;
     private String noticeCode;
     private String idF24;
